Adding Python crossResource
Due to limitations of the PyPi API and not having a central collection of Python Scientific packages (like rOpenSci), I created a list of my own. This is not a definitive list, and others may know useful packages to add to the list.

How it works:
Pull the setup.py file from each package as text from their github repo.  Look for package names in the text. The packages should either be listed as a dependency or listed in a separate requirements text file.

Results are written as a CSV file.